requirements.txt
Adafruit-Blinka==4.3.0
adafruit-circuitpython-neopixel==5.1.1
Adafruit-PlatformDetect==2.5.0
Adafruit-PureIO==1.1.0
click==7.1.1
Flask==1.1.2
Flask-Cors==3.0.8
Flask-SocketIO==4.2.1
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
pkg-resources==0.0.0
pyftdi==0.50.0
pyserial==3.4
python-engineio==3.12.1
python-socketio==4.5.1
pyusb==1.0.2
rpi-ws281x==4.2.3
six==1.14.0
sysv-ipc==1.0.1
Werkzeug==1.0.1
